Extract active record report
