+++ lib/bidu/house/report.rb
@@ -2,6 +2,7 @@ module Bidu
   module House
     class Report
       include JsonParser
+      require 'bidu/house/report/active_record'
       require 'bidu/house/report/error'
       require 'bidu/house/report/range'
       ALLOWED_PARAMETERS = []
@@ -9,8 +10,7 @@ module Bidu
 
       attr_reader :json
 
-      json_parse :period, type: :period
-      json_parse :clazz, :base_scope, :id, case: :snake
+      json_parse :id, case: :snake
 
       def initialize(options)
         @json = DEFAULT_OPTION.merge(options)
@@ -26,26 +26,6 @@ module Bidu
 
       def as_json
         { status: status }
-      end
-
-      private
-
-      def fetch_scoped(base, scope)
-        if (scope.is_a?(Symbol))
-          scope.to_s.split('.').inject(base) do |entries, method|
-            entries.public_send(method)
-          end
-        else
-          base.where(scope)
-        end
-      end
-
-      def last_entries
-        @last_entries ||= base.where('updated_at >= ?', period.seconds.ago)
-      end
-
-      def base
-        fetch_scoped(clazz, base_scope)
       end
     end
   end
